Adding interceptNetworkError to AxiosAuthRefreshOptions
Adds an option to allow intercepting a network error if `options.interceptNetworkError` is `true`.

This is useful for intercepting CORS failures. The Salesforce REST API is an example of an API that supports CORS when authenticated, but doesn't return CORS headers once an access token has expired or is invalid.